Add a vsock server/client example of interaction with C lib #27
build.rs is added to rust-vsock-payload to set location and name of
the C static library. Method to set these information is added to
vsock readme.md.

In vsock_c_lib/main.c, socket interfaces provided by rust is called
to implement a server and a client.

In main.rs of rust-vsock-payload, these two functions implemented by
C lib is called for a test.

Signed-off-by: Jiaqi Gao <jiaqi.gao@intel.com>