rust-vsock-payload: refine the vsock c interfaces.
Refine the vsock Rust-C interfaces.
Update C source code so that it can also be built with clang as
PE format.
Update the build.rs to set a default value of static lib name and
location.
Update the readme.md to show how to compile C source file as
PE/ELF format.

Signed-off-by: Jiaqi Gao <jiaqi.gao@intel.com>